# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação Swagger da API de comandas da prova de suficiência Web II.",
        "title": "REST API Comandas FURB",
        "contact": {
            "name": "Vitor da Silva",
            "email": "vitsilva@furb.br"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/RestAPIFurb/comandas": {
            "get": {
                "description": "Retorna todas as comandas cadastradas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comandas"
                ],
                "summary": "Lista todas as comandas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Comanda"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova comanda com os dados fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comandas"
                ],
                "summary": "Cria nova comanda",
                "parameters": [
                    {
                        "description": "Dados da nova comanda",
                        "name": "comanda",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Comanda"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Comanda"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/RestAPIFurb/comandas/{id}": {
            "get": {
                "description": "Retorna uma comanda pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comandas"
                ],
                "summary": "Busca comanda por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Comanda",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Comanda"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza uma comanda existente pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comandas"
                ],
                "summary": "Atualiza comanda existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Comanda",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualizar a comanda",
                        "name": "comanda",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Comanda"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Comanda"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove uma comanda existente pelo ID (Autenticação JWT necessária)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comandas"
                ],
                "summary": "Remove comanda existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Comanda",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/RestAPIFurb/login": {
            "post": {
                "description": "Gera um token JWT simples para teste",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticação"
                ],
                "summary": "Gera token JWT",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Comanda": {
            "type": "object",
            "required": [
                "idUsuario",
                "nomeUsuario",
                "produtos",
                "telefoneUsuario"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "idUsuario": {
                    "type": "integer"
                },
                "nomeUsuario": {
                    "type": "string"
                },
                "produtos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Produto"
                    }
                },
                "telefoneUsuario": {
                    "type": "string"
                }
            }
        },
        "model.Produto": {
            "type": "object",
            "required": [
                "nome",
                "preco"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}